requirements.txt
# Core dependencies
streamlit>=1.28.0
opencv-python>=4.8.0
mediapipe>=0.10.0
scikit-learn>=1.2.0
numpy>=1.24.0

# AI Translation
google-generativeai>=0.3.0

# Mouse Control
autopy>=4.0.0

# Text-to-speech (optional)
pyttsx3>=2.90

# Environment variables
python-dotenv>=1.0.0
